pyproject.toml
[project]
name = "saviant-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb[cli]>=1.0.13",
    "fastapi>=0.115.13",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.4",
    "langchain-huggingface>=0.3.0",
    "langchain-tavily>=0.2.4",
    "langgraph>=0.4.10",
    "langsmith>=0.4.2",
    "openai>=1.91.0",
    "pandas>=2.3.0",
    "pymupdf>=1.26.1",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.1",
    "python-socketio>=5.13.0",
    "pyyaml>=6.0.2",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.46.0",
    "uvicorn>=0.34.3",
]
